requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
SQLAlchemy==2.0.36
pymysql==1.1.0
python-jose==3.3.0
cryptography==43.0.1
bcrypt==4.0.1